% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{facet_metric}
\alias{facet_metric}
\title{facet_metric}
\usage{
facet_metric(x, y, ttl = NA, yl = NA)
}
\arguments{
\item{x}{month of the year as integer}

\item{y}{acronym of a metric (see `vars.lst`)}

\item{ttl}{}
}
\value{
a ggplot2 plot
}
\description{
Build a yearly-faceted daily plot with averages
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
