% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hvars.lst}
\alias{hvars.lst}
\title{hvars.lst}
\format{
An object of class \code{character} of length 3.
}
\usage{
hvars.lst
}
\description{
List of time recording for metrics in `dataset`
}
\keyword{datasets}
