% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteolecco.R
\docType{package}
\name{meteolecco}
\alias{meteolecco}
\title{meteolecco}
\description{
Historical data about temperature and wind metrics in my small home town Lecco, Italy,
}
\details{
Measures collected by \url{https://www.meteolecco.it}
}
